benchmark/requirements.txt
# Vector DB clients
qdrant-client==1.12.1
weaviate-client==4.9.6
psycopg[binary]==3.2.3
pgvector==0.3.6

# Data + math
numpy==2.1.3
h5py==3.12.1

# Utils
python-dotenv==1.0.1
tqdm==4.67.1
requests==2.32.3
huggingface-hub==0.26.5
pydantic==2.10.3
